+++ Sources/Swift/Integrations/SessionReplay/SentrySessionReplayIntegration.swift
@@ -4,7 +4,7 @@
 #if (os(iOS) || os(tvOS)) && !SENTRY_NO_UI_FRAMEWORK
 import UIKit
 
-typealias SessionReplayIntegrationScope = NotificationCenterProvider & RateLimitsProvider & CurrentDateProvider & RandomProvider & FileManagerProvider & CrashWrapperProvider & ReachabilityProvider & GlobalEventProcessorProvider & DispatchQueueWrapperProvider & ApplicationProvider & DispatchFactoryProvider
+typealias SessionReplayIntegrationScope = NotificationCenterProvider & RateLimitsProvider & CurrentDateProvider & RandomProvider & FileManagerProvider & CrashWrapperProvider & ReachabilityProvider & GlobalEventProcessorProvider & DispatchQueueWrapperProvider & ApplicationProvider & DispatchFactoryProvider & SessionReplayBreadcrumbConverterProvider
 
 // This is static because it will be used for swizzling and would cause retain cycles
 private var touchTracker: SentryTouchTracker?
@@ -27,7 +27,7 @@ public class SentrySessionReplayIntegration: NSObject, SwiftIntegration, SentryS
     private let experimentalOptions: SentryExperimentalOptions
     private let notificationCenter: SentryNSNotificationCenterWrapper
     private var currentScreenshotProvider: SentryViewScreenshotProvider?
-    private var currentBreadcrumbConverter: SentryReplayBreadcrumbConverter?
+    private var breadcrumbConverter: SentryReplayBreadcrumbConverter
     private var previewView: SentryMaskingPreviewView?
     private let dateProvider: SentryCurrentDateProvider
     private let crashWrapper: SentryCrashReporter
@@ -85,6 +85,7 @@ public class SentrySessionReplayIntegration: NSObject, SwiftIntegration, SentryS
         self.random = dependencies.random
         self.crashWrapper = dependencies.crashWrapper
         self.getApplication = dependencies.application
+        self.breadcrumbConverter = dependencies.sessionReplayBreadcrumbConverter
 
         self.replayFileManager = SessionReplayFileManager(
             fileManager: dependencies.fileManager,
@@ -103,7 +104,8 @@ public class SentrySessionReplayIntegration: NSObject, SwiftIntegration, SentryS
             random: random,
             replayProcessingQueue: replayProcessingQueue,
             replayAssetWorkerQueue: replayAssetWorkerQueue,
-            replayFileManager: replayFileManager
+            replayFileManager: replayFileManager,
+            breadcrumbConverter: breadcrumbConverter
         )
         
         setupTouchTrackerIfNeeded(options: options)
@@ -245,7 +247,7 @@ public class SentrySessionReplayIntegration: NSObject, SwiftIntegration, SentryS
             replayOptions: replayOptions,
             experimentalOptions: experimentalOptions,
             screenshotProvider: currentScreenshotProvider ?? viewPhotographer,
-            breadcrumbConverter: currentBreadcrumbConverter ?? SentrySRDefaultBreadcrumbConverter(),
+            breadcrumbConverter: breadcrumbConverter,
             fullSession: startedAsFullSession,
             rootView: window
         )
@@ -272,7 +274,7 @@ public class SentrySessionReplayIntegration: NSObject, SwiftIntegration, SentryS
             replayOptions: replayOptions,
             experimentalOptions: experimentalOptions,
             screenshotProvider: currentScreenshotProvider ?? viewPhotographer,
-            breadcrumbConverter: currentBreadcrumbConverter ?? SentrySRDefaultBreadcrumbConverter(),
+            breadcrumbConverter: breadcrumbConverter,
             fullSession: fullSession,
             rootView: rootView
         )
@@ -399,7 +401,8 @@ public class SentrySessionReplayIntegration: NSObject, SwiftIntegration, SentryS
     @objc public func configureReplayWith(_ breadcrumbConverter: SentryReplayBreadcrumbConverter?, screenshotProvider: SentryViewScreenshotProvider?) {
         SentrySDKLog.debug("[Session Replay] Configuring replay")
         if let bc = breadcrumbConverter {
-            currentBreadcrumbConverter = bc
+            self.breadcrumbConverter = bc
+            replayRecovery?.updateBreadcrumbConverter(bc)
             sessionReplay?.breadcrumbConverter = bc
         }
         if let sp = screenshotProvider {