+++ Sources/Swift/Integrations/SessionReplay/SessionReplayRecovery.swift
@@ -19,19 +19,26 @@ struct SessionReplayRecovery {
     private let replayProcessingQueue: SentryDispatchQueueWrapper
     private let replayAssetWorkerQueue: SentryDispatchQueueWrapper
     private let replayFileManager: SessionReplayFileManager
+    private var breadcrumbConverter: SentryReplayBreadcrumbConverter
     
     init(
         replayOptions: SentryReplayOptions,
         random: SentryRandomProtocol,
         replayProcessingQueue: SentryDispatchQueueWrapper,
         replayAssetWorkerQueue: SentryDispatchQueueWrapper,
-        replayFileManager: SessionReplayFileManager
+        replayFileManager: SessionReplayFileManager,
+        breadcrumbConverter: SentryReplayBreadcrumbConverter
     ) {
         self.replayOptions = replayOptions
         self.random = random
         self.replayProcessingQueue = replayProcessingQueue
         self.replayAssetWorkerQueue = replayAssetWorkerQueue
         self.replayFileManager = replayFileManager
+        self.breadcrumbConverter = breadcrumbConverter
+    }
+
+    mutating func updateBreadcrumbConverter(_ breadcrumbConverter: SentryReplayBreadcrumbConverter) {
+        self.breadcrumbConverter = breadcrumbConverter
     }
     
     // MARK: - Recovery
@@ -167,7 +174,13 @@ struct SessionReplayRecovery {
         var currentSegmentId = config.segmentId
         var currentType = config.type
         for video in videos {
-            captureVideo(video, replayId: replayId, segmentId: currentSegmentId, type: currentType)
+            captureVideo(
+                video,
+                replayId: replayId,
+                segmentId: currentSegmentId,
+                type: currentType,
+                breadcrumbs: event.breadcrumbs ?? []
+            )
             currentSegmentId += 1
             // type buffer is only for the first segment
             currentType = .session
@@ -178,7 +191,13 @@ struct SessionReplayRecovery {
         event.context = eventContext
     }
 
-    private func captureVideo(_ video: SentryVideoInfo, replayId: SentryId, segmentId: Int, type: SentryReplayType) {
+    private func captureVideo(
+        _ video: SentryVideoInfo,
+        replayId: SentryId,
+        segmentId: Int,
+        type: SentryReplayType,
+        breadcrumbs: [Breadcrumb]
+    ) {
         let replayEvent = SentryReplayEvent(
             eventId: replayId,
             replayStartTimestamp: video.start,
@@ -190,7 +209,7 @@ struct SessionReplayRecovery {
         let recording = SentryReplayRecording(
             segmentId: segmentId,
             video: video,
-            extraEvents: []
+            extraEvents: breadcrumbConverter.convert(breadcrumbs, from: video.start, until: video.end)
         )
 
         SentrySDKInternal.currentHub().captureReplayEvent(